Give g_iconv the right kind of pointer.
2004-11-01  Morten Welinder  <terra@gnome.org>

	* ms-excel-read.c (ms_biff_get_chars): Give g_iconv the right kind
	of pointer.

	* ms-excel-write.c (excel_write_string): Ditto.
	(excel_write_string_len): Change second argument to be size_t*.

	* ms-excel-read.h (ms_biff_get_chars): Give g_iconv the right kind
	of pointer.

	* ms-excel-write.h (excel_write_string_len): Change second
	argument to be size_t*.

	* ms-escher.c (ms_escher_get_data): Fix confusion between
	ptrdiff_t and int.

+++ plugins/excel/ms-excel-read.c
@@ -793,10 +793,10 @@ excel_read_string_header (guint8 const *data,
 }
 
 char *
-ms_biff_get_chars (char const *ptr, guint length, gboolean use_utf16)
+ms_biff_get_chars (char const *ptr, size_t length, gboolean use_utf16)
 {
 	char* ans;
-	unsigned i;
+	size_t i;
 
 	if (use_utf16) {
 		gunichar2 *uni_text = g_alloca (sizeof (gunichar2)*length);
